64-bit transition part 10: ksh93 builtin commands
This is the tenth of the thickfold patch series, which enables ksh93
to operate within a 64-bit address space.

The parts of ksh93 affected by this commit are:
- The various primary ksh93 builtins located in the bltins folder.
  - print_namval() returns a value only used for boolean tests,
    so it doesn't need to return nv_size + 1 and thus a warning
    can be quashed.
  - Fixed buggy behavior in the alarm builtin relating to bad
    format specifiers and typing (regression tests also added).
    Output for alarm -r reinput now prints an actual floating point
    and regular 'alarm' reinput now prints the original timeout
    rather than '18446744071934413113.922' or some other unusable
    number.

Change in the number of warnings on Linux when compiling with clang using
-Wsign-compare -Wshorten-64-to-32 -Wsign-conversion -Wimplicit-int-conversion:
911 => 802 => 37 (progression from part 9 => part 10 => part 13)

Progresses https://github.com/ksh93/ksh/issues/592

+++ src/cmd/ksh93/include/builtins.h
@@ -113,7 +113,7 @@ extern int b_times(int, char*[],Shbltin_t*);
     extern int B_echo(int, char*[],Shbltin_t*);
 #endif /* SHOPT_ECHOPRINT */
 
-extern short		b_enum_nelem(Namfun_t*);
+extern ssize_t		b_enum_nelem(Namfun_t*);
 
 extern const char	e_alrm1[];
 extern const char	e_alrm2[];
